refactor: create kraken2 subpackage
