+++ src/taxpasta/infrastructure/application/kraken2/__init__.py
@@ -13,24 +13,8 @@
 # limitations under the License.
 
 
-from .bracken import (
-    BrackenProfile,
-    BrackenProfileReader,
-    BrackenProfileStandardisationService,
-)
-from .centrifuge import (
-    CentrifugeProfile,
-    CentrifugeProfileReader,
-    CentrifugeProfileStandardisationService,
-)
-from .kaiju import KaijuProfile, KaijuProfileReader, KaijuProfileStandardisationService
-from .kraken2 import (
-    Kraken2Profile,
-    Kraken2ProfileReader,
+from .kraken2_profile import Kraken2Profile
+from .kraken2_profile_reader import Kraken2ProfileReader
+from .kraken2_profile_standardisation_service import (
     Kraken2ProfileStandardisationService,
-)
-from .metaphlan_profile import MetaphlanProfile
-from .metaphlan_profile_reader import MetaphlanProfileReader
-from .metaphlan_profile_standardisation_service import (
-    MetaphlanProfileStandardisationService,
 )
