+++ src/taxpasta/infrastructure/application/kraken2/kraken2_profile.py
